backend/requirements.txt
opencv-python
grpcio==1.39.0
grpcio-tools==1.39.0
tensorflow==2.9.1
numpy
